Separate ui components into individual files
