+++ src/samples/map-sample/ol-app/ui/MapTools.tsx
@@ -1,0 +1,91 @@
+// SPDX-FileCopyrightText: 2023 Open Pioneer project (https://github.com/open-pioneer)
+// SPDX-License-Identifier: Apache-2.0
+import { Flex } from "@open-pioneer/chakra-integration";
+import { Geolocation } from "@open-pioneer/geolocation";
+import { InitialExtent, ZoomIn, ZoomOut } from "@open-pioneer/map-navigation";
+import { ToolButton } from "@open-pioneer/react-utils";
+import { useIntl } from "open-pioneer:react-hooks";
+import {
+    PiBookmarksSimpleBold,
+    PiImagesLight,
+    PiListLight,
+    PiMapTrifold,
+    PiRulerLight,
+    PiSelectionPlusBold
+} from "react-icons/pi";
+import { MAP_ID } from "../MapConfigProviderImpl"; // TODO
+
+export interface ToolState {
+    bookmarksActive: boolean;
+    tocActive: boolean;
+    legendActive: boolean;
+    measurementActive: boolean;
+    selectionActive: boolean;
+    overviewMapActive: boolean;
+}
+
+export interface MapToolsProps {
+    toolState: ToolState;
+
+    onToolStateChange(toolStateName: keyof ToolState, newValue: boolean): void;
+}
+
+export function MapTools(props: MapToolsProps) {
+    const intl = useIntl();
+    const { toolState, onToolStateChange } = props;
+
+    const toggleToolState = (name: keyof ToolState) => {
+        onToolStateChange(name, !toolState[name]);
+    };
+
+    return (
+        <Flex
+            role="toolbar"
+            aria-label={intl.formatMessage({ id: "ariaLabel.toolbar" })}
+            direction="column"
+            gap={1}
+            padding={1}
+        >
+            <ToolButton
+                label={intl.formatMessage({ id: "spatialBookmarkTitle" })}
+                icon={<PiBookmarksSimpleBold />}
+                isActive={toolState.bookmarksActive}
+                onClick={() => toggleToolState("bookmarksActive")}
+            />
+            <Geolocation mapId={MAP_ID}></Geolocation>
+            <ToolButton
+                label={intl.formatMessage({ id: "tocTitle" })}
+                icon={<PiListLight />}
+                isActive={toolState.tocActive}
+                onClick={() => toggleToolState("tocActive")}
+            />
+            <ToolButton
+                label={intl.formatMessage({ id: "legendTitle" })}
+                icon={<PiImagesLight />}
+                isActive={toolState.legendActive}
+                onClick={() => toggleToolState("legendActive")}
+            />
+            <ToolButton
+                label={intl.formatMessage({ id: "measurementTitle" })}
+                icon={<PiRulerLight />}
+                isActive={toolState.measurementActive}
+                onClick={() => toggleToolState("measurementActive")}
+            />
+            <ToolButton
+                label={intl.formatMessage({ id: "selectionTitle" })}
+                icon={<PiSelectionPlusBold />}
+                isActive={toolState.selectionActive}
+                onClick={() => toggleToolState("selectionActive")}
+            />
+            <ToolButton
+                label={intl.formatMessage({ id: "overviewMapTitle" })}
+                icon={<PiMapTrifold />}
+                isActive={toolState.overviewMapActive}
+                onClick={() => toggleToolState("overviewMapActive")}
+            />
+            <InitialExtent mapId={MAP_ID} />
+            <ZoomIn mapId={MAP_ID} />
+            <ZoomOut mapId={MAP_ID} />
+        </Flex>
+    );
+}
